# docs/api/openapi.yaml
openapi: 3.0.3
info:
  title: ReMoment Server API
  version: "1"
  description: |
    Client API of ReMoment Server (unauthenticated, local network). Covers what a
    playback/library client needs: server info, devices and playback, radio,
    multiroom, the library, and client registration.

    This file is the contract for native clients (the iOS app is generated from /
    checked against it). The prose docs in `docs/api/*.md` and CLAUDE.md stay the
    detailed reference; when an endpoint's shape changes, change this file too.

    Not yet described here: per-device settings (`sound-adjustment`, `bluetooth`,
    `network`, `wireless-speakers`, `info` — see `docs/api/device-settings.md`) and
    source controls (`/devices/{id}/controls` — see `docs/api/source-controls.md`).

    Errors on device endpoints: `503 unreachable`, `422 unsupported`, `502 driver_error`.
    Always check a device's `capabilities` before calling a feature endpoint.
servers:
  - url: http://remoment.local/api
...
paths:
...
  /devices/{device}/volume:
    parameters: [{ $ref: "#/components/parameters/device" }]
    get:
      tags: [Playback]
      operationId: getVolume
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Volume" }
    put:
      tags: [Playback]
      operationId: setVolume
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/Volume" }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Volume" }
        "422": { $ref: "#/components/responses/DeviceError" }
        "503": { $ref: "#/components/responses/DeviceError" }
...
  /devices/{device}/sources/activate:
    parameters: [{ $ref: "#/components/parameters/device" }]
    post:
      tags: [Devices]
      summary: Switch to a source (capability source_activation)
      operationId: activateSource
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [source_id]
              properties:
                source_id: { type: string }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  status: { type: string }
                  source_id: { type: string }
...
  /devices/{device}/library/play:
    parameters: [{ $ref: "#/components/parameters/device" }]
    post:
      tags: [Library playback]
      summary: Play one library track (capability library_playback)
      operationId: playTrack
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [track_id]
              properties:
                track_id: { type: integer }
      responses:
        "200": { $ref: "#/components/responses/PlayedOk" }
        "422": { $ref: "#/components/responses/DeviceError" }
...
components:
  securitySchemes:
    adminToken:
      type: http
      scheme: bearer
      description: "Token from POST /admin/login (rmt_…). Only routes marked with security need it; the rest of the API is open."
  parameters:
    device:
      name: device
      in: path
      required: true
      schema: { type: integer }
...
  responses:
...
    DeviceError:
      description: unreachable (503), unsupported / not_playable / invalid (422), driver_error (502)
      content:
        application/json:
          schema:
            type: object
            properties:
              error: { type: string, example: unsupported }
              message: { type: string }
    PlayedOk:
      description: OK
      content:
        application/json:
          schema:
            type: object
            properties:
              status: { type: string, example: ok }
              track: { type: string }
              album: { type: string }
              artist: { type: string }
              playlist: { type: string }
  schemas:
...
    Volume:
      type: object
      required: [volume]
      properties:
        volume: { type: integer, minimum: 0, maximum: 100 }
...